add components folder and move all components there
