+++ src/components/Footer/Intervals.jsx
