+++ src/components/Footer/configuration.js
