+++ src/components/Head/Info.jsx
