+++ src/components/Head/constants.js
