+++ src/components/Head/index.js
